Remove unused plotter.py file: duplicated content from log.py
